[refactor] - 패키지 이동 - 이벤트 처리 순서 개선
