+++ src/test/java/kr/hhplus/be/server/api/reservation/integration/ReservationFacadeIntegrationTest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.api.reservation;
+package kr.hhplus.be.server.api.reservation.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -27,7 +27,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 @SpringBootTest
